Trying to do the dodge using booleeans
the dodge will be initiated only if doTheDodge is true, which will be set to false by default, and when the dodge is pressed, it will be set to true and then the body starts moving. After some seconds, it will be returned to false, the setiing true and timing part done by a co routine. So essentially the coroutine just sets the timer and sets the true/false accordingly, and another dodge function which only works if true, will execute outside the co routine, which controls the dodge function using true/false instead of executing it within itself. COMMIT 1 of trial